moved serialization back to utils and made utils.archive classes pydantic models
